+++ crates/config/src/project/mod.rs
@@ -3,9 +3,11 @@ mod global_config;
 mod local_config;
 mod task;
 mod task_options;
+mod workspace;
 
 pub use dep::*;
 pub use global_config::*;
 pub use local_config::*;
 pub use task::*;
 pub use task_options::*;
+pub use workspace::*;
